Renamed folder to qiskit_chemistry_cmd and qiskit_chemistry_ui
