+++ qiskit_chemistry_ui/__main__.py
@@ -20,9 +20,9 @@ import os
 
 qiskit_aqua_chemistry_directory = os.path.dirname(os.path.realpath(__file__))
 qiskit_aqua_chemistry_directory = os.path.join(qiskit_aqua_chemistry_directory, '../..')
-sys.path.insert(0, 'qiskit_aqua_chemistry_ui')
+sys.path.insert(0, 'qiskit_chemistry_ui')
 sys.path.insert(0, qiskit_aqua_chemistry_directory)
 
-from qiskit_aqua_chemistry_ui.command_line import main
+from qiskit_chemistry_ui.command_line import main
 
 main()
