+++ qiskit_chemistry_ui/_sectionpropertiesview.py
