+++ qiskit_chemistry_ui/_sectiontextview.py
