+++ requirements.txt
@@ -1,5 +1,4 @@
 qiskit-aqua>=0.4.0
-qiskit-terra>=0.7.0,<0.8
 numpy>=1.13
 h5py
 psutil>=5
